Add unique label on flags
For now we only have one way to find a Flag: the description.

But description are not unique and very volatile and should not be
used to find a precise Flag imo.

ID aren't good enough too since they're linked to the database and
they change accross Flagr instances (e.g staging/production)

An unique label fixes all of these issues.